improved builing process

+++ create_appimage.sh
@@ -1,9 +1,11 @@
 #! /bin/sh
 
 cd build
-cmake  -DCMAKE_INSTALL_PREFIX=/usr -DEXPERIMENTAL=1 -DENABLE_PYTHON=1 -DENABLE_LIBFIVE=1 ..
+cmake  -DCMAKE_INSTALL_PREFIX=/usr -DEXPERIMENTAL=1 -DENABLE_PYTHON=1 -DPYTHON_VERSION=3.13 -DENABLE_LIBFIVE=1 ..
+rm -rf ../pythonscad.appdata.xml # TODO fix here
 cp ../pythonscad.appdata.xml.in2 ../pythonscad.appdata.xml # TODO fix here
 make -j3
+if [ ! $? == 0 ] ; then exit 1 ; fi
 rm -rf ../AppDir
 make install DESTDIR=../AppDir
 cd ..
